+++ src/bookmark_reader/safari.rs
@@ -1,9 +1,51 @@
+use super::{SelectSource, SourceOs};
 use crate::{bookmarks::SourceBookmarkBuilder, ReadBookmark, Source, SourceBookmarks, SourceType};
 use log::{debug, trace};
 use plist::{Dictionary, Value};
-use std::path::Path;
+use std::path::{Path, PathBuf};
 
 pub type PlistBookmarkReader<'a> = Box<dyn ReadBookmark<'a, ParsedValue = plist::Value>>;
+
+pub struct SafariSelector;
+
+impl SafariSelector {
+    pub fn new() -> Box<Self> {
+        Box::new(SafariSelector)
+    }
+}
+
+impl SelectSource for SafariSelector {
+    fn name(&self) -> SourceType {
+        SourceType::Safari
+    }
+
+    fn source_os(&self) -> SourceOs {
+        SourceOs::Macos
+    }
+
+    fn extension(&self) -> Option<&str> {
+        Some("plist")
+    }
+
+    fn find_sources(&self, home_dir: &Path) -> Result<Vec<PathBuf>, anyhow::Error> {
+        debug!("Find sources for {}", self.name());
+        let browser_dirs = [home_dir.join("Library/Safari")];
+        let bookmark_files = browser_dirs
+            .into_iter()
+            .filter_map(|bookmark_dir| {
+                let bookmark_file = bookmark_dir.join("Bookmarks.plist");
+
+                if bookmark_file.is_file() {
+                    Some(bookmark_file)
+                } else {
+                    None
+                }
+            })
+            .collect::<Vec<_>>();
+
+        Ok(bookmark_files.to_vec())
+    }
+}
 
 /// A bookmark reader to read bookmarks in plist format from Safari.
 #[derive(Debug)]
@@ -137,10 +179,49 @@ mod test {
     use super::*;
     use crate::{
         bookmark_reader::{source_reader::PlistReader, ParsedBookmarks, ReadSource, SourceReader},
-        test_utils, utils,
+        test_utils::{self, tests},
+        utils,
     };
     use assert_matches::assert_matches;
     use std::{collections::HashMap, path::PathBuf};
+    use tempfile::tempdir;
+
+    #[test]
+    fn test_selector_name() {
+        let selector = SafariSelector;
+        assert_eq!(selector.name(), SourceType::Safari);
+    }
+
+    #[test]
+    fn test_find_sources_empty() {
+        let temp_dir = tempdir().unwrap();
+        let temp_path = temp_dir.path();
+        assert!(temp_path.exists(), "Missing path: {}", temp_path.display());
+
+        let selector = SafariSelector;
+        let res = selector.find_sources(temp_path);
+        assert!(res.is_ok(), "{}", res.unwrap_err());
+
+        let sources = res.unwrap();
+        assert!(sources.is_empty());
+    }
+
+    #[test]
+    fn test_find_sources() {
+        let temp_dir = tempdir().unwrap();
+        let temp_path = temp_dir.path();
+        assert!(temp_path.exists(), "Missing path: {}", temp_path.display());
+
+        tests::create_test_files(temp_path);
+
+        let selector = SafariSelector;
+        let res = selector.find_sources(temp_path);
+        assert!(res.is_ok(), "Can't find dir: {}", res.unwrap_err());
+
+        let bookmark_dirs = res.unwrap();
+        assert_eq!(bookmark_dirs.len(), 1);
+        assert!(bookmark_dirs.contains(&temp_path.join("Library/Safari/Bookmarks.plist")));
+    }
 
     #[test]
     fn test_read_and_parse_xml() {